apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.1,<5.2",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
]
